Move make_divisible func
